Remake finish, ready for debugging
Added auto-test feature which tries out nearly any possible income...
finished the code so we can work on debugging
fetch(key,file) = outputs a tuple of keys list and bool (if it was searched for)
unpack(string,file) = outputs a tupla of strings and bool
find(search,inside=langcodes,outputlist=False,isdictionary=False,errorOut=False) = searches for search inside inside, if outputlist, outputs a list, else it outputs the first find, searches in values of a dictionary if isdictionary and errors out when nothing is found and errorOut... if not errorOut, it outputs the same thing as input
google(string, target, source) = the message maker and adder of support for finding keys and putting them to search or smth
path: str = the absolute path to \lang\code.json
debug: bool = the thing that allows you to test googler when offline and also tries every input combination
langcodes: list: str = automatically generated list of strings obtaining the codes of languages used for searching

currently any problems are very likely to be caused by googler, not other functions! be sure to check out the indexing and usage of variables!